refactor: tests/integration/0.command.dev.test.cjs
Replace ava with vitest

- convert file to ESM
- replace ava with vitest
- replace got with node-fetch calls in some tests
